🚚 change directory structure
