+++ src/services/index.ts
@@ -1,5 +1,4 @@
 export * from "./join";
-export * from "./mention";
 export * from "./interaction";
 export * from "./vc";
 export * from "./webhook";
