+++ src/services/interaction/mention.ts
@@ -1,8 +1,8 @@
 import type { Message } from "discord.js";
 import { Locale } from "discord.js";
 
-import { mentionHelpEmbed } from "../embed";
-import { buildEmbed } from "../utils";
+import { mentionHelpEmbed } from "../../embed";
+import { buildEmbed } from "../../utils";
 
 export const mention = async (message: Message) => {
   try {
